python/requirements.txt
flask==3.0.2
flask-cors==4.0.0
RPi.GPIO==0.7.1 